Future proof processor UIDs in MADT parsing
The later x2APIC extended the processor UID to a 4-byte field. We want to
use the same representation for that, so simply use `u32` and cast for the
old single-byte entries.

+++ acpi/src/interrupt.rs
@@ -22,7 +22,9 @@ pub enum LocalInterruptLine {
 #[derive(Debug)]
 pub enum NmiProcessor {
     All,
-    ProcessorUid(u8),
+    /// Refers to a processor with the given UID. This is stored as a `u32`, but should be casted to `u8` when the
+    /// DSDT uses the deprecated `DefProcessor` operator to define processor UIDs.
+    ProcessorUid(u32),
 }
 
 #[derive(Debug)]
